Added separate rotated wheel PNGs for size consistency, removed rotate CSS styling, and added a max width condition for the wheel images of the viewport width
