+++ src/component/LandingPageV2/ProductDevLifecycle/Production/index.jsx
@@ -7,7 +7,7 @@ const Production = () => {
           <>
             <div className="background-black-wheel">
               <div className = "DevelopmentWheel">
-                <img  className="rotate" style={{width: '64vh'}} src ={DevelopmentWheel} />
+                <img style={{width: 'min(64vh, 100vw)'}} src ={DevelopmentWheel} />
               </div>
             </div>
             <div className="background-black-text">
